Rework the offline-plugins integrations tests
