+++ package.json
@@ -10,7 +10,7 @@
     "pretest:unit": "docker-compose up -d",
     "test:unit": "npm run nyc",
     "posttest:unit": "docker-compose stop",
-    "test:integration": "lerna run test",
+    "test:integration": "(cd tests/serverless-plugins-integration && npm test)",
     "test": "npm run eslint && npm run test:unit && npm run test:integration",
     "publish": "lerna publish"
   },
