--- tests/serverless-offline-dynamodb-streams-integration/docker-compose.yml
@@ -1,20 +1,0 @@
-version: "3"
-services:
-  dynamodb:
-    image: cnadiminti/dynamodb-local:latest
-    ports:
-      - "8000:8000"
-  dynamodb-create:
-    image: infrastructureascode/aws-cli:latest
-    links:
-      - dynamodb
-    entrypoint: sh
-    volumes:
-      - ./:/project
-      - ./create-tables.sh:/project/create-tables.sh:ro
-    environment:
-      - AWS_ACCESS_KEY_ID=local
-      - AWS_SECRET_ACCESS_KEY=local
-      - AWS_DEFAULT_REGION=eu-west-1
-      - AWS_ENDPOINT_URL=http://dynamodb:8000
-    command: " ./create-tables.sh"
