--- tests/serverless-offline-kinesis-integration/docker-compose.yml
@@ -1,20 +1,0 @@
-version: "3"
-services:
-  kinesis:
-    image: dlsniper/kinesalite:latest
-    ports:
-      - "4567:4567"
-  kinesis-create:
-    image: infrastructureascode/aws-cli:latest
-    links:
-      - kinesis
-    entrypoint: sh
-    volumes:
-      - ./:/project
-      - ./create-streams.sh:/project/create-streams.sh:ro
-    environment:
-      - AWS_ACCESS_KEY_ID=local
-      - AWS_SECRET_ACCESS_KEY=local
-      - AWS_DEFAULT_REGION=eu-west-1
-      - AWS_ENDPOINT_URL=http://kinesis:4567
-    command: " ./create-streams.sh"
