--- tests/serverless-offline-kinesis-integration/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "serverless-offline-kinesis-integration",
-  "private": true,
-  "version": "0.0.2",
-  "scripts": {
-    "start": "sls offline",
-    "pretest": "docker-compose stop kinesis && docker-compose rm -f kinesis && docker-compose run kinesis-create",
-    "test": "node ./test.js"
-  },
-  "dependencies": {
-    "aws-sdk": "^2.444.0",
-    "figures": "^3.0.0",
-    "lodash": "^4.17.15",
-    "serverless": "^1.41.1",
-    "serverless-offline": "^4.10.0",
-    "serverless-offline-kinesis": "^2.0.0",
-    "signal-exit": "^3.0.2"
-  }
-}
