--- tests/serverless-offline-sqs-integration/handler.js
@@ -1,9 +1,0 @@
-module.exports.promise = (event, context) => {
-  console.log(JSON.stringify(event, null, 4));
-  return Promise.resolve();
-};
-
-module.exports.callback = (event, context, cb) => {
-  console.log(JSON.stringify(event, null, 4));
-  cb();
-};
