--- tests/serverless-offline-sqs-integration/handler.py
@@ -1,5 +1,0 @@
-import json
-
-def handler(event, context):
-    print(event)
-    return
