--- tests/serverless-offline-sqs-integration/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "serverless-offline-sqs-integration",
-  "private": true,
-  "version": "0.0.2",
-  "scripts": {
-    "start": "sls offline",
-    "pretest": "docker-compose stop sqs && docker-compose rm -f sqs && docker-compose run sqs-create",
-    "test": "node ./test.js"
-  },
-  "dependencies": {
-    "aws-sdk": "^2.444.0",
-    "figures": "^3.0.0",
-    "lodash": "^4.17.15",
-    "serverless": "^1.41.1",
-    "serverless-offline": "^4.10.0",
-    "serverless-offline-sqs": "^2.0.3",
-    "signal-exit": "^3.0.2"
-  }
-}
