+++ tests/serverless-plugins-integration/docker-compose.yml
@@ -1,0 +1,59 @@
+version: "3"
+services:
+
+  dynamodb:
+    image: cnadiminti/dynamodb-local:latest
+    ports:
+      - "8000:8000"
+  dynamodb-create:
+    image: infrastructureascode/aws-cli:latest
+    links:
+      - dynamodb
+    entrypoint: sh
+    command: ./create-tables.sh
+    volumes:
+      - ./:/project
+      - ./scripts/create-tables.sh:/project/create-tables.sh:ro
+    environment:
+      - AWS_ACCESS_KEY_ID=local
+      - AWS_SECRET_ACCESS_KEY=local
+      - AWS_DEFAULT_REGION=eu-west-1
+      - AWS_ENDPOINT_URL=http://dynamodb:8000
+
+  sqs:
+    image: s12v/elasticmq:latest
+    ports:
+      - "9324:9324"
+  sqs-create:
+    image: infrastructureascode/aws-cli:latest
+    links:
+      - sqs
+    entrypoint: sh
+    command: ./create-queues.sh
+    volumes:
+      - ./:/project
+      - ./scripts/create-queues.sh:/project/create-queues.sh:ro
+    environment:
+      - AWS_ACCESS_KEY_ID=local
+      - AWS_SECRET_ACCESS_KEY=local
+      - AWS_DEFAULT_REGION=eu-west-1
+      - AWS_ENDPOINT_URL=http://sqs:9324
+
+  kinesis:
+    image: dlsniper/kinesalite:latest
+    ports:
+      - "4567:4567"
+  kinesis-create:
+    image: infrastructureascode/aws-cli:latest
+    links:
+      - kinesis
+    entrypoint: sh
+    command: ./create-streams.sh
+    volumes:
+      - ./:/project
+      - ./scripts/create-streams.sh:/project/create-streams.sh:ro
+    environment:
+      - AWS_ACCESS_KEY_ID=local
+      - AWS_SECRET_ACCESS_KEY=local
+      - AWS_DEFAULT_REGION=eu-west-1
+      - AWS_ENDPOINT_URL=http://kinesis:4567
