+++ tests/serverless-plugins-integration/lambda/handler.js
