+++ tests/serverless-plugins-integration/lambda/handler.py
