+++ tests/serverless-plugins-integration/scripts/create-queues.sh
