+++ tests/serverless-plugins-integration/scripts/create-streams.sh
