+++ tests/serverless-plugins-integration/scripts/create-tables.sh
